python/pyproject.toml
[project]
name = "evolutionary-art-python"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.7.2",
    "scipy>=1.17.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
